refactor: default-image를 포게더용 이미지로 변경
